fix: move cart logic to HOC
