+++ client/routes.js
@@ -10,13 +10,38 @@ import {
   UserProfile
 } from './components'
 import {me} from './store'
+import {reduceOrdersToGetPastCart} from './components/helperFunctions'
+import {fetchMergePastAndGuestCarts} from './store/cart'
 
 /**
  * COMPONENT
  */
 class Routes extends Component {
+  getUserCart() {
+    const unfulfilledOrder = this.props.user.orders.filter(
+      order => !order.date
+    )[0]
+    const pastCart = reduceOrdersToGetPastCart(unfulfilledOrder)
+    this.props.loadCart(pastCart)
+  }
+
   componentDidMount() {
     this.props.loadInitialData()
+    // if we're coming directly from logging in
+    // we WILL have access to user.id in componentDidMount
+    // so we need to make the same call to getUserCart()
+    if (this.props.user.id) this.getUserCart()
+
+    // here we call loadCart with no pastCart
+    // so that if user is NOT logged in
+    // they still get their guest cart from localStorage
+    this.props.loadCart({})
+  }
+
+  // user is not immediately available in componentDidMount
+  // so we load the user's pastCart in componentDidUpdate
+  componentDidUpdate(prevProps) {
+    if (!prevProps.user.id && this.props.user.id) this.getUserCart()
   }
 
   render() {
@@ -42,17 +67,21 @@ const mapState = state => {
   return {
     // Being 'logged in' for our purposes will be defined has having a state.user that has a truthy id.
     // Otherwise, state.user will be an empty object, and state.user.id will be falsey
+    user: state.user,
     isLoggedIn: !!state.user.id
   }
 }
 
-const mapDispatch = dispatch => {
-  return {
-    loadInitialData() {
-      dispatch(me())
-    }
+const mapDispatch = dispatch => ({
+  loadInitialData: () => dispatch(me()),
+
+  // load guestCart from localStorage and pass along with pastCart
+  // to merge the two in the redux store
+  loadCart: pastCart => {
+    let cartFromLocalStorage = JSON.parse(localStorage.getItem('cart')) || {}
+    dispatch(fetchMergePastAndGuestCarts(pastCart, cartFromLocalStorage))
   }
-}
+})
 
 // The `withRouter` wrapper makes sure that updates are not blocked
 // when the url changes
@@ -62,6 +91,5 @@ export default withRouter(connect(mapState, mapDispatch)(Routes))
  * PROP TYPES
  */
 Routes.propTypes = {
-  loadInitialData: PropTypes.func.isRequired,
-  isLoggedIn: PropTypes.bool.isRequired
+  loadInitialData: PropTypes.func.isRequired
 }
